mcp-servers/google-drive/server.py
import json
import logging
import os
from mcp.server.fastmcp import FastMCP
from google.oauth2.credentials import Credentials
from google_auth_oauthlib.flow import InstalledAppFlow
from google.auth.transport.requests import Request
from googleapiclient.discovery import build
from googleapiclient.http import MediaFileUpload

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger("vulcan-gdrive-bridge")

# Initialize FastMCP Server
mcp = FastMCP("vulcan-gdrive-bridge")

# Scopes
SCOPES = ["https://www.googleapis.com/auth/drive"]
CREDENTIALS_FILE = os.path.join(
    os.path.dirname(__file__), "../../secrets/gdrive-credentials.json"
)
TOKEN_FILE = os.path.join(os.path.dirname(__file__), "token.json")


def get_service():
    """Authenticate and return the Drive service."""
    creds = None
    if os.path.exists(TOKEN_FILE):
        creds = Credentials.from_authorized_user_file(TOKEN_FILE, SCOPES)

    if not creds or not creds.valid:
        if creds and creds.expired and creds.refresh_token:
            creds.refresh(Request())
        else:
            if not os.path.exists(CREDENTIALS_FILE):
                raise FileNotFoundError(
                    f"Credentials file not found at {CREDENTIALS_FILE}"
                )

            flow = InstalledAppFlow.from_client_secrets_file(CREDENTIALS_FILE, SCOPES)
            creds = flow.run_local_server(port=0)

        with open(TOKEN_FILE, "w") as token:
            token.write(creds.to_json())

    return build("drive", "v3", credentials=creds)


@mcp.tool()
async def drive_upload(
    local_path: str, remote_folder: str = "Vulcan_Exports", filename: str = None
) -> str:
    """Upload a file to Google Drive."""
    try:
        service = get_service()

        # files = service.files().list(q=f"name='{remote_folder}' and mimeType='application/vnd.google-apps.folder'", spaces='drive').execute()
        # For simplicity, just upload to root or find folder logic here
        # (Simplified implementation for elite speed, assumes folder ID or root for now if search fails)

        file_metadata = {"name": filename or os.path.basename(local_path)}
        media = MediaFileUpload(local_path, resumable=True)

        file = (
            service.files()
            .create(body=file_metadata, media_body=media, fields="id")
            .execute()
        )
        return f"File ID: {file.get('id')}"
    except Exception as e:
        return f"Error: {str(e)}"


@mcp.tool()
async def drive_list(folder: str = None, limit: int = 50) -> str:
    """List files in Google Drive."""
    try:
        service = get_service()
        results = (
            service.files()
            .list(pageSize=limit, fields="nextPageToken, files(id, name)")
            .execute()
        )
        items = results.get("files", [])
        return json.dumps(items, indent=2)
    except Exception as e:
        return f"Error: {str(e)}"


if __name__ == "__main__":
    mcp.run()
